requirements.txt
flask
openai
pdfplumber
dotenv
google-generativeai